Update 2146 - Senha.c

+++ 2146 - Senha.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 
 int main (void){
-    int n;
+    long int n;
     
     while(scanf("%ld", &n) != EOF)
         printf("%ld\n", n-1);
